refactor: decompose XCTestServiceClient into delegate modules
Following the pattern established in #1111 for AccessibilityServiceClient,
this change decomposes XCTestServiceClient (1,435 lines) into focused
delegate modules under src/features/observe/ios/:

- XCTestServiceGestures: tap, swipe, drag, pinch operations
- XCTestServiceText: setText, clearText, imeAction, selectAll
- XCTestServiceHierarchy: hierarchy retrieval, caching, conversion
- XCTestServiceScreenshot: screenshot capture
- XCTestServiceNavigation: pressHome, launchApp

The original file becomes a backward-compatible re-export shim.
Main client reduced from 1,435 to 755 lines (47% reduction).
